mk ReactWebUI, extract styles
